requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
drf-yasg
openpyxl
python-dotenv